Added Happiness face materials. Changed some other materials. Modified the structure of TickManager. Session now owns the Thread object. Changed TickManager to use a lock to prevent overlapping ticks. A tick can now only process if no other is active. (Makes debugging possible aswell) Added starting stuff to DynamicDiffucltyAdjuster.

+++ Assets/0 Scripts/Infrastructure/Residents/Happiness.cs
@@ -89,11 +89,11 @@
             float value = 0;
             if (EvaluateElectricity)
             {
-                if (resident.Home.HasPower) value += 1 / EvaluationQuantity;
+                if (resident.Home.HasPower) value += (float)1 / EvaluationQuantity;
             }
             if (EvaluateWater)
             {
-                if (resident.Home.HasWaterSupply) value += 1 / EvaluationQuantity;
+                if (resident.Home.HasWaterSupply) value += (float)1 / EvaluationQuantity;
             }
             if (EnableBaseHappiness)
             {
